refactor: executeSubscriptionRootField
to operate only on ExecutionContext
